Implement workspace role helpers and enforce access
